Add files via upload
Tom and Jerry is a word game. In which the user placed in the setting of Tom's living room in Tom and Jerry. The objective of the game is to collect all the items in the house to win. The user may look at items, pick up/drop items, and go to different rooms.  The game supports "look", "get", "drop"  commands. The different rooms include the Living Room, Garage, and Kitchen.  The items include a rat trap, cheese, and  dog treats. To enter the different rooms, you must have collected items. For example, to enter the kitchen, you must have grabbed the dog treats from the living room, to enter the garage you, however, do not need any item. The user may find items using the "look" command.